Identified bug in cross-matching that is revealed in the histogram at the end of ml_experimentation.
I have manually followed up in a throw-away python session - the source of the problem is with the values stored in the mpeak column of the histories. The upid values are an exact match. Maybe there is an interpretation problem with the value of mpeak stored in the history data?